# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Get Coffee Cafe
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /user:
    x-swagger-router-controller: user
    get:
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
      description: Get a single user profile
      operationId: getUser
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Register User
      operationId: register
      parameters:
        - name: user
          in: body
          required: true
          description: User's login information
          schema:
            $ref: "#/definitions/RegisterUserBody"
      responses:
        "200":
          description: Successfully registered User
          schema:
            $ref: "#/definitions/UserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    required:
    - message
    properties:
      response-info:
        $ref: "#/definitions/GeneralResponseInfo" 
      message:
        type: string
  GeneralResponseInfo:
    required:
      - success
      - title
      - message
    properties:
      success:
        type: boolean
      title:
        type: string
      message:
        type: string
  GetUserResponse:
    required:
      - user
    properties:
      user:
        $ref: "#/definitions/UserResponse"
  UserResponse:
    properties:
      email:
        type: string
      password:
        type: string
      createdOn:
        type: string
        format: date-time
      response-info:
        $ref: "#/definitions/GeneralResponseInfo"
...
  CredentialBody:
    properties:
      email:
        type: string
      password:
        type: string    
  RegisterUserBody:
    properties:
      user:
        $ref: "#/definitions/CredentialBody"